requirements.txt
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
numpy==1.21.2
requests==2.26.0
python-dotenv==0.19.0
TA-Lib==0.4.24 